homepage-split Split homepage sections into separate components
